routing/requirements.txt
fastapi
uvicorn
python-dotenv
slowapi
prometheus-fastapi-instrumentator
pydantic
